fix cart

+++ src/utils/currency.ts
@@ -1,5 +1,5 @@
 export const lasPlatas = (monto: string) => {
-  return parseFloat(monto).toLocaleString('es-PE', {
+  return parseInt(monto).toLocaleString('es-PE', {
     currency: 'PEN',
     style: 'currency',
   });
